#17 Restructure file structure to match architecture See: https://github.com/df-map-archive/dfma-html5-map-viewer/issues/17
