+++ src/components/read-map-info.js
