+++ src/components/setup-starting-map.js
